requirements.txt
# API + HTTP
requests==2.31.0

# Retry logic
tenacity==8.2.3

# Postgres
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Synthetic data generation
faker==24.0.0

# AWS
boto3==1.34.0

# Snowflake
snowflake-connector-python==3.6.0

# Kafka
confluent-kafka==2.3.0

# dbt
dbt-core==1.7.0
dbt-snowflake==1.7.0

# Airflow
apache-airflow==2.8.0

# Testing
pytest==7.4.0
pytest-mock==3.12.0